trade-data-pro/backend/requirements.txt
# FastAPI 框架
fastapi==0.115.0
uvicorn[standard]==0.32.0

# 数据库
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.0

# 数据验证
pydantic==2.9.2
pydantic-settings==2.6.1

# 爬虫
requests==2.32.3
beautifulsoup4==4.12.3
scrapy==2.12.0
playwright==1.48.0
selenium==4.26.1
fake-useragent==1.5.1

# HTTP 客户端
httpx==0.27.2
aiohttp==3.11.8

# 任务调度
celery==5.4.0
flower==2.0.1

# 数据处理
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5

# 工具
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# 日志
loguru==0.7.2

# 测试
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
